93 - Added main activity test and reorganized folders
